demos/research-interest-in-public-repo/requirements.txt
pyyaml
matplotlib
google-generativeai
python-dotenv